requirements.txt
django==1.11.*
psycopg2-binary==2.8.*
python-dotenv==0.19.2
dj-database-url==0.5.0
